requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# For web interface
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Data processing and visualization
pandas>=2.0.0
plotly>=5.17.0
tabulate>=0.9.0